data_insight_ml/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
matplotlib==3.8.2
seaborn==0.13.0
pyyaml==6.0.1
python-dateutil==2.8.2
joblib==1.3.2
